fix(map): Prevent unnecessary renders/network requests
The way the UI works right now is it mounts/unmounts the map when
viewing the `Data Catalogue` or `Q&A Panel`. This means that when
you switch back, the toolbox hierarchy needs to be re-created, which
isn't good. The `Data Catalogue` and `Q&A Panel` should just overlay
on top of the map, not replace it, reducing the number of requests
being made. The alternative is to use Context, but I feel that is
not what Context is meant for.

IssueID EODHP-307
