+++ src/pages/q-and-a/QAPanel.scss
@@ -1,10 +1,17 @@
 @import '@/styles/main';
 
 .qa-panel {
+  position: absolute;
+  top: 0;
+  right: 0;
   display: flex;
   flex-direction: column;
   height: 100%;
-  padding: 1rem;
+  width: 100%;
+  z-index: 99999;
+
+  // padding: 1rem;
+  background-color: #efefef;
 
   .title {
     font-size: large;
